backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
neo4j==5.16.0
langchain==0.3.15
langchain-openai==0.3.10
langchain-core==0.3.28
python-dotenv==1.0.0
tiktoken==0.5.2
